refactor(#729): restructure endpoints
Internal and admin endpoints were clear named
and checkes were updated.
