+++ .secrets.baseline
@@ -343,5 +343,5 @@
       }
     ]
   },
-  "generated_at": "2025-11-05T21:13:15Z"
+  "generated_at": "2025-11-06T19:47:50Z"
 }
