+++ src/binaries/hanami/src/api/http_endpoints/cluster/create_cluster_v1_0.rs
@@ -189,7 +189,6 @@ async fn check_quota(context: &UserContext) -> Result<(), ErrorResponse> {
     let max_number_of_meta_clusters = match get_quota(
         miko_endpoint,
         &context.token,
-        &config::CONFIG.api.internal_api_key,
         &context.user_id,
         config::CONFIG.insecure_clients,
     )
