+++ src/binaries/miko/src/api/http_endpoints/quota/get_quota_admin_v1_0.rs
@@ -12,64 +12,58 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use actix_web::web::Json;
 use actix_web::web::Path;
-use apistos::actix::NoContent;
 use apistos::api_operation;
 
 use crate::database::quota_table;
-use crate::database::user_table;
 
 use ainari_api::errors::ErrorResponse;
+use ainari_api_structs::quota_structs::*;
 use ainari_api_structs::user_context::UserContext;
 use ainari_common::enums;
 
 #[api_operation(
-    tag = "user",
-    summary = "Delete user",
-    description = r###"Delete a user from the database. This can only be done by an admin."###,
+    tag = "quota",
+    summary = "Get quota",
+    description = r###"Get information of a quota from the database."###,
     error_code = 400,
     error_code = 401,
     error_code = 404,
     error_code = 500
 )]
-pub async fn delete_user(
+pub async fn get_quota_admin(
     user_id: Path<String>,
     context: UserContext,
-) -> Result<NoContent, ErrorResponse> {
+) -> Result<Json<QuotaResp>, ErrorResponse> {
     if !context.is_admin {
         return Err(ErrorResponse::Unauthorized(
             "Only Admins are allowed to use this endpoint".to_string(),
         ));
     }
 
-    if context.user_id == user_id.to_string() {
-        return Err(ErrorResponse::Conflict(
-            "A user can not delete himself.".to_string(),
-        ));
-    }
+    match quota_table::get_quota(&user_id, &context) {
+        Ok(quota) => {
+            let resp = QuotaResp {
+                user_id: quota.id.clone(),
+                max_cluster: quota.max_cluster,
+                max_dataset: quota.max_dataset,
+                max_checkpoint: quota.max_checkpoint,
+                max_secret: quota.max_secret,
+                max_taskqueue: quota.max_taskqueue,
+                created_by: quota.created_by.clone(),
+                created_at: quota.created_at.clone(),
+                updated_by: quota.updated_by.clone(),
+                updated_at: quota.updated_at.clone(),
+            };
 
-    // delete quota of user from database
-    match quota_table::delete_quota(&user_id, &context) {
-        Ok(_) => {}
+            return Ok(Json(resp));
+        }
         Err(enums::DbError::InternalError) => {
             return Err(ErrorResponse::InternalError("".to_string()));
         }
         Err(enums::DbError::NotFound) => {
             let msg = format!("Quota of user with ID '{user_id}' not found.");
-            return Err(ErrorResponse::NotFound(msg));
-        }
-    };
-
-    // delete user from database
-    match user_table::delete_user(&user_id, &context) {
-        Ok(_) => {
-            return Ok(NoContent);
-        }
-        Err(enums::DbError::InternalError) => {
-            return Err(ErrorResponse::InternalError("".to_string()));
-        }
-        Err(enums::DbError::NotFound) => {
-            let msg = format!("User with ID '{user_id}' not found.");
             return Err(ErrorResponse::NotFound(msg));
         }
     };
