+++ src/binaries/miko/src/api/http_endpoints/user/mod.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod create_user_v1_0;
-pub mod delete_user_v1_0;
-pub mod get_user_v1_0;
-pub mod list_user_v1_0;
+pub mod create_user_admin_v1_0;
+pub mod delete_user_admin_v1_0;
+pub mod get_user_admin_v1_0;
+pub mod list_user_admin_v1_0;
