+++ src/binaries/omamori/src/api/http_endpoints/secret/create_secret_v1_0.rs
@@ -117,7 +117,6 @@ async fn check_quota(context: &UserContext) -> Result<(), ErrorResponse> {
     let max_number_of_secrets = match get_quota(
         miko_endpoint,
         &context.token,
-        &config::CONFIG.api.internal_api_key,
         &context.user_id,
         config::CONFIG.insecure_clients,
     )
