+++ src/binaries/sakura/src/api/http_endpoints/cluster/mod.rs
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod create_cluster_v1_0;
-pub mod delete_cluster_v1_0;
-pub mod get_cluster_v1_0;
-pub mod list_cluster_v1_0;
+pub mod create_cluster_internal_v1_0;
+pub mod delete_cluster_internal_v1_0;
+pub mod get_cluster_internal_v1_0;
+pub mod list_cluster_internal_v1_0;
 pub mod request_cluster_v1_0;
 pub mod train_cluster_v1_0;
