+++ src/binaries/torii/src/api/routes/v1alpha.rs
@@ -33,7 +33,7 @@ pub fn v1alpha_routes() -> Scope {
                 .service(resource("/{proxy_uuid}").route(get().to(get_proxy_v1_0::get_proxy)))
                 .service(
                     resource("/{proxy_uuid}/internal")
-                        .route(delete().to(delete_proxy_v1_0::delete_proxy)),
+                        .route(delete().to(delete_proxy_internal_v1_0::delete_proxy_internal)),
                 ),
         )
 }
