+++ src/libs/rust/ainari_clients/src/quota.rs
@@ -15,7 +15,6 @@
 use ainari_api_structs::quota_structs::*;
 use ainari_common::config as ainari_config;
 use ainari_common::error::AinariError;
-use ainari_common::secret::Secret;
 
 use crate::handle_response;
 use crate::prepare_client;
@@ -23,18 +22,16 @@ use crate::prepare_client;
 pub async fn get_quota(
     miko_endpoint: &ainari_config::MikoEndpoint,
     token: &String,
-    internal_api_key: &Secret,
-    user_id: &String,
+    user_id: &str,
     insecure_client: bool,
 ) -> Result<QuotaResp, AinariError> {
     let address = miko_endpoint.address.clone();
     let client = prepare_client(&address, insecure_client);
-    let url = format!("{address}/v1alpha/quota/{user_id}");
+    let url = format!("{address}/v1alpha/quota");
 
     let response = client
         .get(url)
         .insert_header(("Authorization", format!("Bearer {}", token)))
-        .insert_header(("X-Internal-API-Key", internal_api_key.reveal()))
         .send()
         .await;
